pyproject.toml
[project]
name = "saw-ring-ino"
version = "0.1.0"
description = ""
authors = [
    {name = "koushi hiraoka",email = "hiraoka.koushi@arakawa-lab.com"}
]
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "pyserial (>=3.5,<4.0)",
    "numpy (>=2.3.1,<3.0.0)",
    "matplotlib (>=3.10.3,<4.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
